'.db' database file extension was removed
There's no reason to force the file extension. It should be up to
the developer to provide the desired file name with or without file
extension. Moreover, forcing file extension may prevent sessions to
be store in a single database unless the main database file also
uses the '.db' file extension.